move eventrecorder to handle/eventrecorder
