+++ guest/handle/eventrecorder/api/eventrecorder_test.go
@@ -18,7 +18,7 @@ package api_test
 
 import (
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/api/proto"
-	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/eventrecorder/api"
+	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/handle/eventrecorder/api"
 )
 
 var (
