+++ guest/handle/eventrecorder/eventrecorder.go
@@ -21,8 +21,8 @@ import (
 	"runtime"
 
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/api/proto"
-	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/eventrecorder/api"
-	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/eventrecorder/internal"
+	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/handle/eventrecorder/api"
+	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/handle/eventrecorder/internal"
 	"sigs.k8s.io/kube-scheduler-wasm-extension/guest/internal/mem"
 )
